29/07/2024 (add blacklist filter)

+++ BigMarket-domain/src/main/java/org/example/domain/strategy/model/entity/RuleActionEntity.java
@@ -32,6 +32,6 @@ public class RuleActionEntity <T extends RuleActionEntity.RaffleEntity>{
 
         private String ruleWeightValueKey;
 
-        private Integer awardId;
+        private Long awardId;
     }
 }
